break out more types
