+++ client/session.go
@@ -23,7 +23,6 @@ package client
 import (
 	"errors"
 	"fmt"
-	"io"
 	"math"
 	"sort"
 	"sync"
@@ -31,15 +30,12 @@ import (
 	"time"
 
 	"github.com/m3db/m3db/clock"
-	"github.com/m3db/m3db/context"
 	"github.com/m3db/m3db/encoding"
 	"github.com/m3db/m3db/generated/thrift/rpc"
 	"github.com/m3db/m3db/network/server/tchannelthrift/convert"
-	"github.com/m3db/m3db/storage/block"
 	"github.com/m3db/m3db/storage/bootstrap"
 	"github.com/m3db/m3db/topology"
 	"github.com/m3db/m3db/ts"
-	xio "github.com/m3db/m3db/x/io"
 
 	xerrors "github.com/m3db/m3x/errors"
 	xlog "github.com/m3db/m3x/log"
@@ -1751,316 +1747,6 @@ func (s *session) streamBlocksReattemptFromPeers(
 	})
 }
 
-type blocksResult struct {
-	sync.RWMutex
-	opts                    Options
-	blockOpts               block.Options
-	blockAllocSize          int
-	contextPool             context.Pool
-	encoderPool             encoding.EncoderPool
-	multiReaderIteratorPool encoding.MultiReaderIteratorPool
-	bytesPool               pool.BytesPool
-	result                  bootstrap.ShardResult
-}
-
-func newBlocksResult(
-	opts Options,
-	bootstrapOpts bootstrap.Options,
-	multiReaderIteratorPool encoding.MultiReaderIteratorPool,
-) *blocksResult {
-	blockOpts := bootstrapOpts.DatabaseBlockOptions()
-	return &blocksResult{
-		opts:                    opts,
-		blockOpts:               blockOpts,
-		blockAllocSize:          blockOpts.DatabaseBlockAllocSize(),
-		contextPool:             opts.ContextPool(),
-		encoderPool:             blockOpts.EncoderPool(),
-		multiReaderIteratorPool: multiReaderIteratorPool,
-		bytesPool:               blockOpts.BytesPool(),
-		result:                  bootstrap.NewShardResult(4096, bootstrapOpts),
-	}
-}
-
-func (r *blocksResult) addBlockFromPeer(id ts.ID, block *rpc.Block) error {
-	var (
-		start    = time.Unix(0, block.Start)
-		segments = block.Segments
-		result   = r.blockOpts.DatabaseBlockPool().Get()
-	)
-
-	if segments == nil {
-		return errSessionBadBlockResultFromPeer
-	}
-
-	switch {
-	case segments.Merged != nil:
-		// Unmerged, can insert directly into a single block
-		size := len(segments.Merged.Head) + len(segments.Merged.Tail)
-		data := r.bytesPool.Get(size)[:size]
-		n := copy(data, segments.Merged.Head)
-		copy(data[n:], segments.Merged.Tail)
-
-		encoder := r.encoderPool.Get()
-		encoder.ResetSetData(start, data, false)
-
-		result.Reset(start, encoder)
-
-	case segments.Unmerged != nil:
-		// Must merge to provide a single block
-		readers := make([]io.Reader, len(segments.Unmerged))
-		for i := range segments.Unmerged {
-			readers[i] = xio.NewSegmentReader(ts.Segment{
-				Head: segments.Unmerged[i].Head,
-				Tail: segments.Unmerged[i].Tail,
-			})
-		}
-		encoder, err := r.mergeReaders(start, readers)
-		if err != nil {
-			return err
-		}
-		result.Reset(start, encoder)
-
-	default:
-		return errSessionBadBlockResultFromPeer
-
-	}
-
-	// No longer need the encoder, seal the block
-	result.Seal()
-
-	resultCtx := r.contextPool.Get()
-	defer resultCtx.Close()
-
-	resultReader, err := result.Stream(resultCtx)
-	if err != nil {
-		return err
-	}
-	if resultReader == nil {
-		return nil
-	}
-
-	var tmpCtx context.Context
-
-	for {
-		r.Lock()
-		currBlock, exists := r.result.BlockAt(id, start)
-		if !exists {
-			r.result.AddBlock(id, result)
-			r.Unlock()
-			break
-		}
-
-		// Remove the existing block from the result so it doesn't get
-		// merged again
-		r.result.RemoveBlockAt(id, start)
-		r.Unlock()
-
-		// If we've already received data for this block, merge them
-		// with the new block if possible
-		if tmpCtx == nil {
-			tmpCtx = r.contextPool.Get()
-		} else {
-			tmpCtx.Reset()
-		}
-
-		currReader, err := currBlock.Stream(tmpCtx)
-		if err != nil {
-			return err
-		}
-
-		// If there are no data in the current block, there is no
-		// need to merge
-		if currReader == nil {
-			continue
-		}
-
-		readers := []io.Reader{currReader, resultReader}
-		encoder, err := r.mergeReaders(start, readers)
-		tmpCtx.BlockingClose()
-
-		if err != nil {
-			return err
-		}
-
-		result.Reset(start, encoder)
-		result.Seal()
-	}
-
-	return nil
-}
-
-func (r *blocksResult) mergeReaders(start time.Time, readers []io.Reader) (encoding.Encoder, error) {
-	iter := r.multiReaderIteratorPool.Get()
-	iter.Reset(readers)
-	defer iter.Close()
-
-	encoder := r.encoderPool.Get()
-	encoder.Reset(start, r.blockAllocSize)
-
-	for iter.Next() {
-		dp, unit, annotation := iter.Current()
-		if err := encoder.Encode(dp, unit, annotation); err != nil {
-			encoder.Close()
-			return nil, err
-		}
-	}
-	if err := iter.Err(); err != nil {
-		encoder.Close()
-		return nil, err
-	}
-
-	return encoder, nil
-}
-
-type receivedBlocks struct {
-	submitted bool
-	results   []*blocksMetadata
-}
-
-type processFn func(batch []*blocksMetadata)
-
-type peerBlocksQueue struct {
-	sync.RWMutex
-	closed       bool
-	peer         hostQueue
-	queue        []*blocksMetadata
-	doneFns      []func()
-	assigned     uint64
-	completed    uint64
-	maxQueueSize int
-	workers      xsync.WorkerPool
-	processFn    processFn
-}
-
-type newPeerBlocksQueueFn func(
-	peer hostQueue,
-	maxQueueSize int,
-	interval time.Duration,
-	workers xsync.WorkerPool,
-	processFn processFn,
-) *peerBlocksQueue
-
-func newPeerBlocksQueue(
-	peer hostQueue,
-	maxQueueSize int,
-	interval time.Duration,
-	workers xsync.WorkerPool,
-	processFn processFn,
-) *peerBlocksQueue {
-	q := &peerBlocksQueue{
-		peer:         peer,
-		maxQueueSize: maxQueueSize,
-		workers:      workers,
-		processFn:    processFn,
-	}
-	if interval > 0 {
-		go q.drainEvery(interval)
-	}
-	return q
-}
-
-func (q *peerBlocksQueue) drainEvery(interval time.Duration) {
-	for {
-		q.Lock()
-		if q.closed {
-			q.Unlock()
-			return
-		}
-		q.drainWithLock()
-		q.Unlock()
-		time.Sleep(interval)
-	}
-}
-
-func (q *peerBlocksQueue) close() {
-	q.Lock()
-	q.closed = true
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) trackAssigned(amount int) {
-	atomic.AddUint64(&q.assigned, uint64(amount))
-}
-
-func (q *peerBlocksQueue) trackCompleted(amount int) {
-	atomic.AddUint64(&q.completed, uint64(amount))
-}
-
-func (q *peerBlocksQueue) enqueue(bm *blocksMetadata, doneFn func()) {
-	q.Lock()
-
-	if len(q.queue) == 0 && cap(q.queue) < q.maxQueueSize {
-		// Lazy initialize queue
-		q.queue = make([]*blocksMetadata, 0, q.maxQueueSize)
-	}
-	if len(q.doneFns) == 0 && cap(q.doneFns) < q.maxQueueSize {
-		// Lazy initialize doneFns
-		q.doneFns = make([]func(), 0, q.maxQueueSize)
-	}
-	q.queue = append(q.queue, bm)
-	if doneFn != nil {
-		q.doneFns = append(q.doneFns, doneFn)
-	}
-	q.trackAssigned(len(bm.blocks))
-
-	// Determine if should drain immediately
-	if len(q.queue) < q.maxQueueSize {
-		// Require more to fill up block
-		q.Unlock()
-		return
-	}
-	q.drainWithLock()
-
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) drain() {
-	q.Lock()
-	q.drainWithLock()
-	q.Unlock()
-}
-
-func (q *peerBlocksQueue) drainWithLock() {
-	if len(q.queue) == 0 {
-		// None to drain
-		return
-	}
-	enqueued := q.queue
-	doneFns := q.doneFns
-	q.queue = nil
-	q.doneFns = nil
-	q.workers.Go(func() {
-		q.processFn(enqueued)
-		// Call done callbacks
-		for i := range doneFns {
-			doneFns[i]()
-		}
-		// Track completed blocks
-		completed := 0
-		for i := range enqueued {
-			completed += len(enqueued[i].blocks)
-		}
-		q.trackCompleted(completed)
-	})
-}
-
-type peerBlocksQueues []*peerBlocksQueue
-
-func (qs peerBlocksQueues) findQueue(peer hostQueue) *peerBlocksQueue {
-	for _, q := range qs {
-		if q.peer == peer {
-			return q
-		}
-	}
-	return nil
-}
-
-func (qs peerBlocksQueues) closeAll() {
-	for _, q := range qs {
-		q.close()
-	}
-}
-
 func newTimesByUnixNanos(values []int64) []time.Time {
 	result := make([]time.Time, len(values))
 	for i := range values {
@@ -2075,54 +1761,4 @@ func newTimesByRPCBlocks(values []*rpc.Block) []time.Time {
 		result[i] = time.Unix(0, values[i].Start)
 	}
 	return result
-}
-
-type metadataIter struct {
-	inputCh  <-chan blocksMetadata
-	errCh    <-chan error
-	host     topology.Host
-	blocks   []block.Metadata
-	metadata block.BlocksMetadata
-	done     bool
-	err      error
-}
-
-func newMetadataIter(inputCh <-chan blocksMetadata, errCh <-chan error) PeerBlocksMetadataIter {
-	return &metadataIter{
-		inputCh: inputCh,
-		errCh:   errCh,
-		blocks:  make([]block.Metadata, 0, blocksMetadataInitialCapacity),
-	}
-}
-
-func (it *metadataIter) Next() bool {
-	if it.done || it.err != nil {
-		return false
-	}
-	m, more := <-it.inputCh
-	if !more {
-		it.err = <-it.errCh
-		it.done = true
-		return false
-	}
-	it.host = m.peer.Host()
-	var zeroed block.Metadata
-	for i := range it.blocks {
-		it.blocks[i] = zeroed
-	}
-	it.blocks = it.blocks[:0]
-	for _, b := range m.blocks {
-		bm := block.NewMetadata(b.start, b.size, b.checksum)
-		it.blocks = append(it.blocks, bm)
-	}
-	it.metadata = block.NewBlocksMetadata(m.id, it.blocks)
-	return true
-}
-
-func (it *metadataIter) Current() (topology.Host, block.BlocksMetadata) {
-	return it.host, it.metadata
-}
-
-func (it *metadataIter) Err() error {
-	return it.err
 }
